Added feature: Action Icon
Now an action icon is displayed on top of the gladiators to indicate the action that the gladiators are taking. When the turn is resolving, the user can see also the enemys gladiators action. Before that the user will only see his own action plus an action taken confirmation from the enemy