New battle balance + bug fixing
Added a function improveEnemysGladiator() that balances the battle difficulty. Now the enemis will be a bit stronger by inheriting stats from either the hero himself, or a previously defeated enemy.

Fixed a bug that would make the hero focused when retrieved from the DB for first time. (Attribute not validated as a number and was being passed as string by mistake)

Increased the size of the battle log

Increased the size of the battle background