Several updates. Bug fixing, small new implementations
Refined AI of the enemy. Now they don't always attack when the hero is on low HP; they have a chance of defending as well

Now, when the hero isdead it is not possible to have a combat. We can only observe the battle layout but not fighting

Created a CSS folder in the project directory for better organization

Minor visual improvements -> The Action icon has now a shadow. -> The focused text has now a shadow

Fixed minor bug that was showing the enemys action icon as a loading effect before the battle had even started

The HP bars are now dynamic. Every time a gladiator loses HP, the bars graphics on screen get updated accordingly

Fixed a logic issue that created a bug where the Enemy was not losing the focused state when focused after defending a hit by the focused hero